Move everything from NJTAIM
